requirements.txt
Pillow==9.4.0
streamlit==1.16.0
